Make clock position configurable

+++ connection.cpp
@@ -183,10 +183,10 @@ void updateLocalDataFromServer() {
     Serial.println(F("Day data from server collected"));
 }
 
-bool waitForNTPUpdate(uint32_t ms_timeout)
+bool waitForNTPUpdate(unsigned long ms_timeout)
 {
     static constexpr time_t START_OF_YEAR_2024 = 1704067200;
-    uint32_t start = millis();
+    unsigned long start = millis();
     while((millis()-start) <= ms_timeout) {
         time_t now = time(nullptr);
