+++ connection.hpp
@@ -6,6 +6,6 @@
 void connectToWiFi();
 void configNTP();
 void updateLocalDataFromServer();
-bool waitForNTPUpdate(uint32_t timeout_ms = 5000);
+bool waitForNTPUpdate(unsigned long timeout_ms = 5000);
 
 #endif  // RWCLOCK_CONNECTION_HPP_
